+++ PCSMenu/src/Functions/Scripts/checkLinux.sh
