+++ PCSMenu/src/main.rs
@@ -1,0 +1,6 @@
+mod menu;
+mod utils;
+
+fn main() {
+    menu::main();
+}
